backend/requirements.txt
asgiref==3.10.0
dj-database-url==1.2.0
Django==4.0
django-nested-admin==4.0.1
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==25.0
psycopg2-binary==2.9.9
python-monkey-business==1.1.0
pytz==2025.2
sqlparse==0.5.3
whitenoise==6.5.0
